+++ Platform Specific/Thrower.UnitTests.NETSTD13/Properties/AssemblyInfo.cs
@@ -2,6 +2,8 @@
 //
 // Author(s): Alessio Parma <alessio.parma@gmail.com>
 //
+// The MIT License (MIT)
+//
 // Copyright (c) 2013-2016 Alessio Parma <alessio.parma@gmail.com>
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
@@ -19,20 +21,17 @@
 // DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT
 // OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using System;
 using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
-using System.Security;
 
 // General Information about an assembly is controlled through the following set of attributes.
 // Change these attribute values to modify the information associated with an assembly.
-[assembly: AssemblyTitle("PommaLabs.Thrower")]
-[assembly: AssemblyDescription("Fully managed library providing convenience methods to perform argument checks.")]
+[assembly: AssemblyTitle("PommaLabs.Thrower.UnitTests")]
+[assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("PommaLabs")]
-[assembly: AssemblyProduct("PommaLabs.Thrower")]
-[assembly: AssemblyCopyright("Copyright (c) 2013-2016 Alessio Parma <alessio.parma@gmail.com>")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("PommaLabs.Thrower.UnitTests")]
+[assembly: AssemblyCopyright("Copyright © 2015")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -46,15 +45,5 @@ using System.Security;
 //
 // You can specify all the values or you can default the Build and Revision Numbers by using the '*'
 // as shown below: [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("4.0.0")]
-[assembly: AssemblyFileVersion("4.0.5")]
-
-// Common Language Specification (CLS) compliance generally refers to the claim that CLS rules and
-// restrictions are being followed.
-[assembly: CLSCompliant(true)]
-
-// Specifies that an assembly cannot cause an elevation of privilege.
-[assembly: SecurityTransparent]
-
-// Added to improve unit testing.
-[assembly: InternalsVisibleTo("PommaLabs.Thrower.UnitTests, PublicKey=0024000004800000940000000602000000240000525341310004000001000100efe22a8787b348e219d6b501b425d79f31502681d5a9f7d11030852e44e7ef2b29ddbc7dfcd6461fc3e67a6d7a186dea40535dad461679209a079a45440a99440bd292498b623e9a7fc5161c519c8d45b79fdfbd95a6e0ac5e211d5438d47b8635662d108448cb84bf1212983fab0010c68b41d5168bcbe5a59f9149974852c4")]
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]