+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * SerializableAttribute is no more defined as a polyfill for .NET Standard 1.3; instead, it is taken from System.Runtime.Serialization.Formatters package.
 * Added new DLL for .NET Standard 1.2.
+* Fixed a security setting on .NET Standard 1.3 which prevented FastMember to work properly.
 
 ### v4.0.5 (2016-12-25) ###